♻️ move getTestReportDirectory to envUtils
