--- test/getTestReportDirectory.js
@@ -1,5 +1,0 @@
-module.exports = function getTestReportDirectory() {
-  if (process.env.CI_JOB_NAME) {
-    return `test-report/${process.env.CI_JOB_NAME}`
-  }
-}
